fix(coordinator): Update connections compatibility check to check version instead of namespace
The check for database connections was comparing the ident and the namespace to the ident and version,
which can never be true.

Also related to this bug, the `userConfig` defined in capabilities from reactr is always definied by default.
As such, the check for `userConfig.DB != nil` always evaluated to true, overwriting the database definied
in connections.

Fixes: #166